perf(queue): dispatch the four consumers without importing the API graph
The queue dispatcher reads batch.queue and hands off to one of four handlers,
each behind its own dynamic import. It needed none of the route graph -- it just
lived in server/index.ts, which is where the default export lives, so reaching it
evaluated all 426 routes and every Zod schema first. It now lives in
server/queue.ts, and workers/app.ts imports that directly. server/index.ts still
exports the same shape.

Moving the dispatcher alone would have achieved nothing: it called
handleSyncDlqBatch, and that lived in portal/integration.module.ts, which
statically imports integration.routes and statutory-admin.routes. Reaching the
DLQ handler through it dragged two route modules back in. The handler is now
server/portal/sync-dlq.ts -- it needs OutboxService and a logger, nothing else --
and integration.module re-exports it, so server/index.ts and both test files that
import it are unchanged.

WARNING, and this is the part worth reading: the analogous cron split measured
273ms -> 59ms on a COLD isolate locally and changed production by NOTHING.
Measured 2026-09-07 over ~40min post-deploy: 10.48ms/tick before, 10.86ms after.
Production invocations land on isolates other traffic has already warmed, so the
import was already amortised there and cost nothing to remove. A queue consumer
is invoked the same way. The honest expectation is a cheaper cold start that is
paid rarely -- not a per-invocation win. Both the module and the entry say so
where the next person will read them, rather than repeating the claim.

test:workers 36 files / 181 tests green.
